2020-09-09-0814-v1 (#292)
* renamed tests from #271

* configure student_web as single page flask app (#255)

* catchall routes with flask app (closes #255)

* build: release-2020-08-22-v1

* unittest complete eventlog/index.html in progress (#274)

* create event log view page (close #274)

* change LOG_TYPE to IntEnum. Remove alert (close #274)

* create paging (close #274)

* create generic partial view for pager (#280)

* add paging to lessons (closes #280), check lesson on learningobjectives index view (closes #272), add html5 to index templates (#264)

* create consistent ui with additional templates and surround components with widget class (closes #264)

* create navigation widgets widget--sidebar (closes #280)

* collapse whole of sidenav (closes #280)

* create progress bar (closes #283), create bread crumb (closes #284)

* lower case url (closes #286)

* create layer for progress bar (closes #286)

* fix saving/upsert keyword, includes scheme_of_work_id, edit keywords at scheme of work level (closes #290)

* new items added to key words for scheme of work can be removed no quibble (closes #290)

* create build for static folder

+++ src/database/schema/storedprocedures/keyword__delete.sql
@@ -4,10 +4,25 @@ DROP PROCEDURE IF EXISTS keyword__delete;
 
 CREATE PROCEDURE keyword__delete (
  IN p_keyword_id INT,
+ IN p_scheme_of_work_id INT,
  IN p_auth_user INT)
 BEGIN
+    INSERT INTO sow_key_word__archive (id, name, definition, scheme_of_work_id)
+    SELECT id, name, definition, scheme_of_work_id
+    FROM sow_key_word
+    WHERE id = p_keyword_id and scheme_of_work_id = p_scheme_of_work_id;
+
+    INSERT INTO sow_lesson__has__key_words__archive (key_word_id, lesson_id)
+    SELECT lkw.key_word_id, lkw.lesson_id
+    FROM sow_key_word kw 
+        INNER JOIN sow_lesson__has__key_words lkw ON lkw.key_word_id = kw.id
+        WHERE kw.id = p_keyword_id and kw.scheme_of_work_id = p_scheme_of_work_id;
+
     DELETE FROM sow_key_word
-    WHERE id = p_keyword_id;
+    WHERE id = p_keyword_id and scheme_of_work_id = p_scheme_of_work_id
+              AND p_auth_user IN (SELECT auth_user_id 
+                                  FROM sow_teacher 
+                                  WHERE auth_user_id = p_auth_user AND scheme_of_work_id = p_scheme_of_work_id);
 END;
 //
